Add the index view for the involuntary conversion task list
We want to be able to view the voluntary task list defined in the
`Conversion::Involuntary::TaskList` model.

We have added a base `TaskListController` in the previous commit that defines
behaviour shared between the task lists.

This adds the index view for the involuntary conversion task list. It:
- Adds a `Conversion::Voluntary::TaskListsController` and index view.
- Moves common tests into shared specs
- For now, while we replace the previous task list, the new task list will be
  found at `projects/conversion/involuntary/:project_id/tasks`
