+++ app/controllers/conversion/involuntary/task_lists_controller.rb
@@ -1,0 +1,2 @@
+class Conversion::Involuntary::TaskListsController < TaskListController
+end
