+++ config/routes.rb
@@ -27,6 +27,7 @@ Rails.application.routes.draw do
       namespace :involuntary do
         get "new", to: "projects#new"
         post "new", to: "projects#create"
+        get ":project_id/tasks", to: "task_lists#index", as: :tasks
       end
 
       namespace :voluntary do
