+++ spec/requests/conversion/voluntary/task_lists_controller_spec.rb
@@ -1,26 +1,9 @@
 require "rails_helper"
 
 RSpec.describe Conversion::Voluntary::TaskListsController do
-  let(:user) { create(:user, :caseworker) }
+  let(:project) { create(:conversion_project) }
+  let(:project_id) { project.id }
+  let(:index_path) { conversion_voluntary_tasks_path(project_id) }
 
-  before do
-    sign_in_with(user)
-    mock_successful_api_responses(urn: 123456, ukprn: 10061021)
-  end
-
-  describe "#index" do
-    let(:project) { create(:conversion_project) }
-    let(:project_id) { project.id }
-
-    subject(:perform_request) do
-      get conversion_voluntary_tasks_path(project_id)
-      response
-    end
-
-    it "returns a successful response" do
-      expect(subject).to have_http_status :success
-
-      expect(response.body).to include "Task list"
-    end
-  end
+  it_behaves_like "a task lists controller"
 end
